backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
sqlalchemy[asyncio]==2.0.36
asyncpg>=0.30.0
psycopg2-binary>=2.9.9
pydantic==2.10.4
pydantic-settings==2.7.1
httpx==0.28.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-multipart==0.0.20
websockets==14.1
redis[hiredis]>=5.0.0
scikit-learn>=1.5.0
numpy>=1.26.0
psutil>=5.9.0

# Honeypots & network
paramiko>=3.4.0
aiohttp>=3.9.0

# Scheduling
APScheduler>=3.10.0

# Vector DB & embeddings
qdrant-client>=1.7.0
sentence-transformers>=2.3.0

# DNS & packet analysis
dnspython>=2.5.0
scapy>=2.5.0

# File monitoring
watchdog>=4.0.0

# Reports
reportlab>=4.1.0

# Database migrations
alembic>=1.13.0

# Testing
pytest>=8.0.0
pytest-asyncio>=0.24.0
httpx==0.28.1
aiosqlite>=0.20.0

# MongoDB Atlas
motor>=3.7.0
pymongo>=4.16.0
